refactor: rename extractor files
