+++ cognite/neat/graph/extractors/__init__.py
@@ -1,11 +1,11 @@
 from cognite.neat.graph.extractors import rdf_to_graph
-from cognite.neat.graph.extractors.graph_sheet_to_graph import (
+from cognite.neat.graph.extractors._graph_capturing_sheet import (
     extract_graph_from_sheet,
     read_graph_excel_file_to_table_by_name,
 )
 
-from .graph_sheet_to_graph import GraphCapturingSheet
-from .mocks.graph import MockGraphGenerator
+from ._graph_capturing_sheet import GraphCapturingSheet
+from .mocks._mock_graph_generator import MockGraphGenerator
 
 __all__ = [
     "MockGraphGenerator",
